e2e-tests reward points: code review suggestions
